refactor: 메인뷰컨트롤러내의 함수 extension으로 분리
